+++ test/Experimental/RAGTools/runtests.jl
@@ -1,6 +1,9 @@
 using Test
 using SparseArrays, LinearAlgebra
 using PromptingTools.Experimental.RAGTools
+using PromptingTools
+using AbstractTrees
+const PT = PromptingTools
 using JSON3, HTTP
 
 @testset "RAGTools" begin
@@ -9,5 +12,6 @@ using JSON3, HTTP
     include("preparation.jl")
     include("retrieval.jl")
     include("generation.jl")
+    include("annotation.jl")
     include("evaluation.jl")
 end